readme

+++ concesionaria/api_v1/views/usuario.py
@@ -3,12 +3,12 @@ from rest_framework import status
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.permissions import IsAuthenticated
 from api_v1.serializer.usuario_serializer import UserSerializer
-from vehiculos.models import User
+from django.contrib.auth.models import User  
 
 class UserViewSet(ModelViewSet):
     queryset = User.objects.all()
     serializer_class = UserSerializer
-    permission_classes = [IsAuthenticated] 
+    permission_classes = [IsAuthenticated]
 
     def create(self, request, *args, **kwargs):
         if not request.user.is_staff:
@@ -18,18 +18,26 @@ class UserViewSet(ModelViewSet):
             )
 
         serializer = self.get_serializer(data=request.data)
+        
         if serializer.is_valid():
             user = serializer.save()
-           
-            password=serializer.validate_data.get('password')
+
+            password = serializer.validated_data.get('password')
             if password:
-               user.ser_password(password)
-               user.save()
+                user.set_password(password)
+                user.save()
+
             return Response(UserSerializer(user).data, status=status.HTTP_201_CREATED)
+        
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    # delete
     def destroy(self, request, *args, **kwargs):
+        if not request.user.is_staff:
+            return Response(
+                {"detail": "No tienes permiso para eliminar usuarios"},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
         instance = self.get_object()
         instance.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
